Former-commit-id: 41325fe5d2806692c53ca30a2b726c695b064b66
Former-commit-id: 0138df79cf3a3bec3ea0d913e4be510e8e5f4588

+++ NINA/Utility/CameraSDKs/FLISDK/FLICameraSDK.cs
@@ -267,7 +267,7 @@ namespace FLI {
         public static extern unsafe uint FLIGetCoolerPower(uint dev, ref double power);
 
         [DllImport(DLLNAME, EntryPoint = "FLIGrabRow", CharSet = CharSet.Ansi, CallingConvention = CallingConvention.StdCall)]
-        public static extern unsafe uint FLIGrabRow(uint dev, [Out] byte[] buff, int width);
+        public static extern unsafe uint FLIGrabRow(uint dev, [Out] ushort[] buff, int width);
 
         [DllImport(DLLNAME, EntryPoint = "FLIExposeFrame", CharSet = CharSet.Ansi, CallingConvention = CallingConvention.StdCall)]
         public static extern unsafe uint FLIExposeFrame(uint dev);
